refactor(schedule): extract REST routes into lib/routes/jobs.ts
Final cut of the schedule god-file split. Moves the 12
defineRoute/searchRoute entries and their zod schemas out of index.ts
into lib/routes/jobs.ts, exported as `scheduleRoutes` and spread into
definePlugin (`routes: scheduleRoutes`).

Pure behavior-preserving relocation: handlers, schemas, and route paths
are unchanged. index.ts sheds 18 now-unused imports and becomes a
183-line definePlugin shell (imports + activate orchestration +
onReady/onShutdown + __scheduleTestInternals). 1,443 -> 183 overall
across the six-phase split (87% reduction); index/schedule is the last
audit god-file resolved.

The route<->exec-tool dedup, pause-`pauseUntil` drift fix, dead settings
keys, and ctx-threading are behavior-touching follow-ups, deliberately
NOT taken here.

Verified: typecheck, eslint (changed files), schedule suite 280-0
(--isolate), full suite 5124-0, binary build (stamps reverted), and a
boot smoke against an openclaw PATH stub (schedule activates; GET / and
POST /parse both 200).
