+++ plugins/schedule/lib/routes/jobs.ts
@@ -1,0 +1,437 @@
+/**
+ * Schedule REST routes — the declarative job CRUD/run/parse/search surface.
+ *
+ * The 12 defineRoute/searchRoute entries plus their zod schemas. Handlers stay
+ * thin: validate → call a job-service verb / the fire engine / the cron parser
+ * → log. Spread into definePlugin from index.ts.
+ */
+import { z } from 'zod'
+import { defineRoute, searchRoute } from '@bakin/core/routing'
+import type { BakinJobMeta } from '../../types'
+import { getJob, upsertJob, newScheduleId, withDefaults } from '../sidecar'
+import { parseSchedule } from '../cron-parser'
+import { getSystemTimezone, json } from '../schedule-util'
+import { checkSchedulePrompt } from '../prompt-guard'
+import { getLastRun, readRuns } from '../runs-reader'
+import { getRuntimeMainAgentId } from '@bakin/core/adapters/runtime'
+import { fireManualRun } from '../fire-engine'
+import { guardBakinMutation, deleteScheduleJob, indexJob, readMergedRuntimeJobs } from '../job-service'
+
+// ─── Schemas ─────────────────────────────────────────────────────────────
+
+const jobIdParams = z.object({ jobId: z.string().min(1) })
+const okResponse = z.object({ ok: z.literal(true) }).passthrough()
+const errorResponse = z.object({ error: z.string() }).passthrough()
+const passthrough = z.object({}).passthrough()
+
+const createJobBody = z.object({
+  name: z.string().min(1),
+  schedule: z.string().min(1),
+  agentId: z.string().optional(),
+  workflowId: z.string().optional(),
+  taskPrompt: z.string().optional(),
+  taskTitle: z.string().optional(),
+  owner: z.string().optional(),
+  requireTriage: z.boolean().optional(),
+  allowOverlap: z.boolean().optional(),
+  maxFailures: z.number().optional(),
+  tz: z.string().optional(),
+})
+
+const updateJobBody = z.object({
+  jobId: z.string().optional(),
+}).passthrough()
+
+const deleteJobBody = z.object({
+  jobId: z.string().optional(),
+}).passthrough().optional()
+
+const pauseJobBody = z.object({
+  jobId: z.string().optional(),
+  action: z.enum(['pause', 'resume', 'skip']),
+  pauseUntil: z.string().optional(),
+  skipN: z.number().optional(),
+})
+
+const runNowBody = z.object({
+  jobId: z.string().optional(),
+}).passthrough().optional()
+
+const runsHistoryQuery = z.object({
+  limit: z.coerce.number().int().positive().optional(),
+})
+
+const parseBody = z.object({
+  input: z.string().min(1),
+})
+
+const adoptJobBody = z.object({
+  jobId: z.string().optional(),
+  name: z.string().optional(),
+  schedule: z.string().optional(),
+  agentId: z.string().nullable().optional(),
+  workflowId: z.string().nullable().optional(),
+  taskPrompt: z.string().nullable().optional(),
+  taskTitle: z.string().nullable().optional(),
+  owner: z.string().nullable().optional(),
+  requireTriage: z.boolean().optional(),
+  allowOverlap: z.boolean().optional(),
+  maxFailures: z.number().optional(),
+  tz: z.string().optional(),
+}).passthrough()
+
+const restoreNativeBody = z.object({
+  jobId: z.string().optional(),
+}).passthrough().optional()
+
+export const scheduleRoutes = [
+  defineRoute({
+    path: '/',
+    method: 'GET',
+    summary: 'List all scheduled jobs',
+    description: 'Returns the merged runtime cron + Bakin sidecar view of all jobs.',
+    responses: { 200: passthrough, 500: errorResponse },
+    handler: async () => {
+      const jobs = (await readMergedRuntimeJobs()).map(j => ({
+        ...j,
+        cron: j.schedule.type === 'cron' ? j.schedule.value : undefined,
+      }))
+      return json({ jobs })
+    },
+  }),
+
+  defineRoute({
+    path: '/',
+    method: 'POST',
+    summary: 'Create a scheduled job',
+    body: createJobBody,
+    responses: { 200: passthrough, 400: errorResponse, 500: errorResponse },
+    handler: async (_req, ctx, { body }) => {
+      const parsed = parseSchedule(body.schedule)
+      if (!parsed) {
+        return json({ error: 'Could not parse schedule expression' }, 400)
+      }
+      const tz = body.tz || getSystemTimezone()
+      const jobId = newScheduleId()
+      const owner = body.owner ?? await getRuntimeMainAgentId(ctx.runtime)
+      const meta: BakinJobMeta = {
+        jobId,
+        isBakinJob: true,
+        source: 'bakin',
+        schedule: { kind: 'cron', expr: parsed.cron },
+        enabled: true,
+        displayName: body.name,
+        agentId: body.agentId,
+        owner,
+        requireTriage: body.requireTriage ?? false,
+        workflowId: body.workflowId,
+        taskPrompt: body.taskPrompt,
+        taskTitle: body.taskTitle,
+        allowOverlap: body.allowOverlap ?? false,
+        maxFailures: body.maxFailures ?? 3,
+        consecutiveFailures: 0,
+        tz,
+        createdAt: new Date().toISOString(),
+        updatedAt: new Date().toISOString(),
+      }
+      upsertJob(meta)
+      indexJob(jobId)
+      ctx.activity.audit('job.created', body.owner ?? 'system', { jobId, name: body.name })
+      ctx.activity.log(body.owner ?? 'system', `Created schedule "${body.name}"`)
+      const warnings = checkSchedulePrompt(body.taskPrompt)
+      return json({ ok: true, jobId, cron: parsed.cron, human: parsed.human, tz, ...(warnings.length ? { warnings } : {}) })
+    },
+  }),
+
+  defineRoute({
+    path: '/:jobId',
+    method: 'PUT',
+    summary: 'Update a scheduled job',
+    params: jobIdParams,
+    body: updateJobBody,
+    responses: { 200: okResponse, 400: errorResponse, 403: errorResponse, 404: errorResponse, 500: errorResponse },
+    handler: async (_req, ctx, { params, body }) => {
+      const jobId = params.jobId || (body as { jobId?: string }).jobId
+      if (!jobId) return json({ error: 'jobId required' }, 400)
+      const guard = await guardBakinMutation(jobId)
+      if (!guard.ok) return json({ error: guard.error }, guard.status)
+      const meta = guard.meta
+      const b = body as Record<string, unknown>
+      if (b.schedule && typeof b.schedule === 'string') {
+        const parsed = parseSchedule(b.schedule)
+        if (!parsed) return json({ error: 'Could not parse schedule' }, 400)
+        meta.schedule = { kind: 'cron', expr: parsed.cron }
+      }
+      if (b.name && typeof b.name === 'string') {
+        meta.displayName = b.name
+      }
+      const sidecarFields = [
+        'displayName', 'description', 'agentId', 'owner', 'requireTriage',
+        'workflowId', 'taskPrompt', 'taskTitle', 'allowOverlap', 'maxFailures',
+      ]
+      for (const field of sidecarFields) {
+        if (b[field] !== undefined) {
+          ;(meta as unknown as Record<string, unknown>)[field] = b[field]
+        }
+      }
+      upsertJob(meta)
+      indexJob(jobId)
+      ctx.activity.audit('job.updated', 'system', { jobId })
+      ctx.activity.log('system', `Updated schedule "${meta.displayName || jobId}"`)
+      const warnings = checkSchedulePrompt(meta.taskPrompt)
+      return json({ ok: true, ...(warnings.length ? { warnings } : {}) })
+    },
+  }),
+
+  defineRoute({
+    path: '/:jobId',
+    method: 'GET',
+    summary: 'Get scheduled job details',
+    params: jobIdParams,
+    responses: { 200: passthrough, 404: errorResponse },
+    handler: async (_req, ctx, { params }) => {
+      const meta = getJob(params.jobId)
+      if (!meta) return json({ error: 'Job not found' }, 404)
+      const defaults = withDefaults(meta, await getRuntimeMainAgentId(ctx.runtime))
+      const lastRun = await getLastRun(ctx.runtime.cron, params.jobId)
+      return json({
+        job: {
+          id: meta.jobId,
+          name: meta.displayName,
+          agent: meta.agentId,
+          owner: defaults.owner,
+          paused: meta.paused ?? false,
+          pauseReason: meta.pauseReason,
+          pauseUntil: meta.pauseUntil,
+          workflowId: meta.workflowId,
+          taskPrompt: meta.taskPrompt,
+          taskTitle: meta.taskTitle,
+          allowOverlap: defaults.allowOverlap,
+          maxFailures: defaults.maxFailures,
+          consecutiveFailures: meta.consecutiveFailures ?? 0,
+          lastTaskId: meta.lastTaskId,
+          lastRun: lastRun ?? null,
+          tz: meta.tz,
+          createdAt: meta.createdAt,
+        },
+      })
+    },
+  }),
+
+  defineRoute({
+    path: '/:jobId',
+    method: 'DELETE',
+    summary: 'Delete a scheduled job',
+    params: jobIdParams,
+    body: deleteJobBody,
+    responses: { 200: okResponse, 400: errorResponse, 403: errorResponse },
+    handler: async (_req, ctx, { params }) => {
+      const jobId = params.jobId
+      const guard = await guardBakinMutation(jobId)
+      if (!guard.ok) return json({ error: guard.error }, guard.status)
+      const result = await deleteScheduleJob(ctx, jobId)
+      ctx.activity.audit('job.deleted', 'system', { jobId })
+      ctx.activity.log('system', `Deleted schedule "${jobId}"`)
+      return json({ ok: true, ...result })
+    },
+  }),
+
+  defineRoute({
+    path: '/:jobId/adopt',
+    method: 'POST',
+    summary: 'Adopt a runtime cron job into Bakin task scheduling',
+    params: jobIdParams,
+    body: adoptJobBody,
+    responses: { 200: passthrough, 400: errorResponse, 404: errorResponse, 409: errorResponse, 500: errorResponse },
+    handler: async (_req, ctx, { params, body }) => {
+      const jobId = params.jobId || body.jobId
+      if (!jobId) return json({ error: 'jobId required' }, 400)
+
+      const runtimeJob = await ctx.runtime.cron.get(jobId)
+      if (!runtimeJob) return json({ error: 'Runtime cron job not found' }, 404)
+
+      const existing = getJob(jobId)
+      if (existing?.isBakinJob) return json({ error: 'Job is already managed by Bakin' }, 409)
+
+      const raw = await ctx.runtime.cron.getRaw(jobId, 'schedule adopt: preserve native cron before Bakin takes ownership')
+      if (!raw) return json({ error: 'Runtime cron snapshot not found' }, 404)
+
+      const parsed = body.schedule ? parseSchedule(body.schedule) : { cron: runtimeJob.schedule }
+      if (!parsed) return json({ error: 'Could not parse schedule' }, 400)
+
+      const now = new Date().toISOString()
+      const tz = body.tz || existing?.tz || getSystemTimezone()
+      const displayName = body.name || existing?.displayName || runtimeJob.name
+      const owner = (body.owner ?? undefined) || existing?.owner || await getRuntimeMainAgentId(ctx.runtime)
+      const taskPrompt = (body.taskPrompt ?? undefined) || existing?.taskPrompt || runtimeJob.command
+
+      const meta: BakinJobMeta = {
+        ...(existing ?? {}),
+        jobId,
+        isBakinJob: true,
+        source: 'adopted',
+        schedule: { kind: 'cron', expr: parsed.cron },
+        enabled: runtimeJob.enabled ?? true,
+        displayName,
+        agentId: body.agentId === null ? undefined : body.agentId ?? existing?.agentId,
+        owner,
+        requireTriage: body.requireTriage ?? existing?.requireTriage ?? false,
+        workflowId: body.workflowId === null ? undefined : body.workflowId ?? existing?.workflowId,
+        taskPrompt,
+        taskTitle: body.taskTitle === null ? undefined : body.taskTitle ?? existing?.taskTitle,
+        allowOverlap: body.allowOverlap ?? existing?.allowOverlap ?? false,
+        maxFailures: body.maxFailures ?? existing?.maxFailures ?? 3,
+        consecutiveFailures: existing?.consecutiveFailures ?? 0,
+        tz,
+        originalRuntimeCron: {
+          provider: ctx.runtime.name,
+          capturedAt: now,
+          snapshot: raw,
+        },
+        createdAt: existing?.createdAt ?? now,
+        updatedAt: now,
+      }
+      upsertJob(meta)
+      // Bakin owns the schedule now — remove the native cron so it stops firing.
+      await ctx.runtime.cron.remove(jobId)
+      indexJob(jobId)
+
+      ctx.activity.audit('job.adopted', 'system', { jobId })
+      ctx.activity.log('system', `Adopted runtime cron "${displayName}" into Bakin`)
+      return json({ ok: true, jobId })
+    },
+  }),
+
+  defineRoute({
+    path: '/:jobId/restore-native',
+    method: 'POST',
+    summary: 'Restore an adopted Bakin schedule back to its native runtime cron behavior',
+    params: jobIdParams,
+    body: restoreNativeBody,
+    responses: { 200: passthrough, 400: errorResponse, 404: errorResponse, 500: errorResponse },
+    handler: async (_req, ctx, { params }) => {
+      const jobId = params.jobId
+      if (!jobId) return json({ error: 'jobId required' }, 400)
+
+      const meta = getJob(jobId)
+      if (!meta?.originalRuntimeCron) return json({ error: 'No native cron snapshot available' }, 404)
+
+      const restored = await ctx.runtime.cron.restoreRaw(
+        jobId,
+        meta.originalRuntimeCron.snapshot,
+        'schedule restore: return adopted cron to native runtime behavior',
+      )
+      const now = new Date().toISOString()
+      const runtimeMeta: BakinJobMeta = {
+        jobId,
+        isBakinJob: false,
+        source: 'runtime',
+        displayName: restored.name,
+        owner: meta.owner ?? await getRuntimeMainAgentId(ctx.runtime),
+        requireTriage: true,
+        createdAt: meta.createdAt ?? now,
+        updatedAt: now,
+      }
+      upsertJob(runtimeMeta)
+      indexJob(jobId)
+
+      ctx.activity.audit('job.restored_native', 'system', { jobId })
+      ctx.activity.log('system', `Restored schedule "${restored.name}" to native runtime cron behavior`)
+      return json({ ok: true, jobId })
+    },
+  }),
+
+  defineRoute({
+    path: '/:jobId/pause',
+    method: 'POST',
+    summary: 'Pause/resume/skip a scheduled job',
+    params: jobIdParams,
+    body: pauseJobBody,
+    responses: { 200: okResponse, 400: errorResponse, 403: errorResponse, 404: errorResponse },
+    handler: async (_req, ctx, { params, body }) => {
+      const jobId = params.jobId || body.jobId
+      if (!jobId) return json({ error: 'jobId required' }, 400)
+      const guard = await guardBakinMutation(jobId)
+      if (!guard.ok) return json({ error: guard.error }, guard.status)
+      const meta = guard.meta
+      switch (body.action) {
+        case 'pause':
+          meta.paused = true
+          meta.pauseReason = 'manual'
+          meta.pauseUntil = body.pauseUntil
+          meta.enabled = false
+          break
+        case 'resume':
+          meta.paused = false
+          meta.pauseReason = undefined
+          meta.pauseUntil = undefined
+          meta.skipNextN = undefined
+          meta.skippedCount = undefined
+          meta.consecutiveFailures = 0
+          meta.enabled = true
+          break
+        case 'skip':
+          if (!meta.isBakinJob) return json({ error: 'Skip is only available for Bakin schedules' }, 400)
+          meta.skipNextN = body.skipN ?? 1
+          meta.skippedCount = 0
+          break
+      }
+      upsertJob(meta)
+      indexJob(jobId)
+      ctx.activity.audit(`job.${body.action}`, 'system', { jobId })
+      ctx.activity.log('system', `Schedule "${meta.displayName || jobId}" ${body.action}d`)
+      return json({ ok: true })
+    },
+  }),
+
+  defineRoute({
+    path: '/:jobId/run',
+    method: 'POST',
+    summary: 'Trigger immediate run',
+    params: jobIdParams,
+    body: runNowBody,
+    responses: { 200: okResponse, 400: errorResponse, 404: errorResponse },
+    handler: async (_req, ctx, { params }) => {
+      const jobId = params.jobId
+      const meta = getJob(jobId)
+      if (!meta?.isBakinJob) return json({ error: 'Job not found' }, 404)
+      await fireManualRun(meta, jobId)
+      ctx.activity.audit('job.run_now', 'system', { jobId })
+      ctx.activity.log('system', `Triggered immediate run for "${jobId}"`)
+      return json({ ok: true })
+    },
+  }),
+
+  defineRoute({
+    path: '/:jobId/runs',
+    method: 'GET',
+    summary: 'Get run history for a job',
+    params: jobIdParams,
+    query: runsHistoryQuery,
+    responses: { 200: passthrough },
+    handler: async (_req, ctx, { params, query }) => {
+      const limit = query.limit ?? 50
+      const runs = await readRuns(ctx.runtime.cron, params.jobId, limit)
+      return json({ runs })
+    },
+  }),
+
+  defineRoute({
+    path: '/parse',
+    method: 'POST',
+    summary: 'Parse a schedule expression',
+    description: 'Converts a natural-language or cron expression into structured schedule output.',
+    body: parseBody,
+    responses: { 200: passthrough, 400: errorResponse },
+    handler: async (_req, _ctx, { body }) => {
+      const result = parseSchedule(body.input)
+      if (!result) {
+        return json({ error: 'Could not parse schedule. Try a simpler expression or raw cron.' }, 400)
+      }
+      return json(result)
+    },
+  }),
+
+
+  searchRoute({ table: 'schedule' }),
+]
